// packages/cactus-plugin-ledger-connector-fabric/src/main/typescript/generated/protos/corda/view_data_pb.ts
// Copyright IBM Corp. All Rights Reserved.
//
// SPDX-License-Identifier: Apache-2.0

// @generated by protoc-gen-es v2.2.2 with parameter "target=ts"
// @generated from file corda/view_data.proto (package corda, syntax proto3)
/* eslint-disable */

import type { GenFile, GenMessage } from "@bufbuild/protobuf/codegenv1";
import { fileDesc, messageDesc } from "@bufbuild/protobuf/codegenv1";
import type { Message } from "@bufbuild/protobuf";

/**
 * Describes the file corda/view_data.proto.
 */
export const file_corda_view_data: GenFile = /*@__PURE__*/
  fileDesc("ChVjb3JkYS92aWV3X2RhdGEucHJvdG8SBWNvcmRhIqEBCghWaWV3RGF0YRI8ChJub3Rhcml6ZWRfcGF5bG9hZHMYASADKAsyIC5jb3JkYS5WaWV3RGF0YS5Ob3Rhcml6ZWRQYXlsb2FkGlcKEE5vdGFyaXplZFBheWxvYWQSEQoJc2lnbmF0dXJlGAEgASgJEhMKC2NlcnRpZmljYXRlGAIgASgJEgoKAmlkGAMgASgJEg8KB3BheWxvYWQYBCABKAxCcAopb3JnLmh5cGVybGVkZ2VyLmNhY3RpLndlYXZlci5wcm90b3MuY29yZGFaQ2dpdGh1Yi5jb20vaHlwZXJsZWRnZXItY2FjdGkvY2FjdGkvd2VhdmVyL2NvbW1vbi9wcm90b3MtZ28vdjIvY29yZGFiBnByb3RvMw");

/**
 * @generated from message corda.ViewData
 */
export type ViewData = Message<"corda.ViewData"> & {
  /**
   * @generated from field: repeated corda.ViewData.NotarizedPayload notarized_payloads = 1;
   */
  notarizedPayloads: ViewData_NotarizedPayload[];
};

/**
 * Describes the message corda.ViewData.
 * Use `create(ViewDataSchema)` to create a new message.
 */
export const ViewDataSchema: GenMessage<ViewData> = /*@__PURE__*/
  messageDesc(file_corda_view_data, 0);

/**
 * @generated from message corda.ViewData.NotarizedPayload
 */
export type ViewData_NotarizedPayload = Message<"corda.ViewData.NotarizedPayload"> & {
  /**
   * @generated from field: string signature = 1;
   */
  signature: string;

  /**
   * @generated from field: string certificate = 2;
   */
  certificate: string;

  /**
   * @generated from field: string id = 3;
   */
  id: string;

  /**
   * Bytes of InteropPayload
   *
   * @generated from field: bytes payload = 4;
   */
  payload: Uint8Array;
};

/**
 * Describes the message corda.ViewData.NotarizedPayload.
 * Use `create(ViewData_NotarizedPayloadSchema)` to create a new message.
 */
export const ViewData_NotarizedPayloadSchema: GenMessage<ViewData_NotarizedPayload> = /*@__PURE__*/
  messageDesc(file_corda_view_data, 0, 0);